Finalizing some features

+++ src/users/services.py
@@ -1,31 +1,31 @@
+import datetime
 import secrets
-import smtplib
 import uuid
+import random
 import jwt
 import logging
 
-from fastapi import Depends, Request
+from fastapi import Depends, Request, UploadFile
 from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
 from fastapi.responses import RedirectResponse
 from datetime import timedelta
 from urllib.parse import urlencode
 from typing import Optional, List
 
-from src.users.models import User, OAuthProvider, Plans, Feedback
+from src.users.models import User, OAuthProvider, Plans, Feedback, CodeType
 from src.users.repositories import UserRepository
 from src.users.schemas import UserCreate, TokenData, UserLogin, FeedbackCreate
 from src.users.exceptions import CredentialException, TokenTypeException, UserNotFoundException, AccessException, \
     EmailExistsException, IncorrectEmailAddressException, IncorrectVerifyCodeException, EmailSenderException, \
-    OAuthServiceNotFoundException, InvalidOAuthStateException
+    OAuthServiceNotFoundException, InvalidOAuthStateException, CodeTypeException, TooLargeFileException
 
 from statistic.schemas import UserDocument, DayStatistic
 from statistic.utils import get_or_create_user
-from tasks.celery_worker import task_send_to_email, task_send_feedback
+from tasks.celery_worker import task_send_to_email
 
-from config_data.constants import welcome_messages
+from config_data.constants import constants, messages
 from config_data.config import Config, load_config
 
-from utils.email_sender import generate_confirmation_mode
 from utils.jwt_settings import validate_password, decode_jwt, encode_jwt
 from utils.oauth2_settings import get_google_oauth_email, get_yandex_oauth_email, get_github_oauth_email
 
@@ -37,6 +37,7 @@ logger = logging.getLogger(__name__)
 settings: Config = load_config(".env")
 auth_config = settings.authJWT
 variables = settings.variablesData
+email_sender = settings.email_sender
 
 TOKEN_TYPE_FIELD = "type"
 ACCESS_TOKEN_TYPE = "access"
@@ -45,6 +46,10 @@ REFRESH_TOKEN_TYPE = "refresh"
 
 class UserService:
     repository = UserRepository()
+
+    @staticmethod
+    def generate_confirmation_mode():
+        return random.randint(email_sender.MIN_CODE, email_sender.MAX_CODE)
 
     async def get_oauth2_redirect(self, request: Request, service: OAuthProvider) -> RedirectResponse:
         service_data = service.value
@@ -134,38 +139,82 @@ class UserService:
 
         return response
 
-    async def send_feedback(self, new_feedback: FeedbackCreate, user: User) -> Feedback:
-        task_send_feedback.delay(new_feedback.name, new_feedback.message, user.email)
+    async def send_feedback(
+            self, new_feedback: FeedbackCreate, user: User, file: Optional[UploadFile] = None
+    ) -> Feedback:
+        if file and file.size > constants.MAX_FEEDBACK_FILE_SIZE:
+            raise TooLargeFileException(file.size, constants.MAX_FEEDBACK_FILE_SIZE)
+
+        message = messages.LETTER_FEEDBACK_MESSAGE.format(from_email=user.email, message=new_feedback.message)
+
+        try:
+            task_send_to_email.delay(
+                subject=messages.LETTER_FEEDBACK_TITLE.format(name=new_feedback.name),
+                body=message,
+                address=email_sender.ADMIN_EMAIL,
+                file_content=file.file.read() if file else None,
+                file_name=file.filename if file else None
+            )
+        except Exception:
+            raise EmailSenderException()
+
+        if file:
+            file.file.close()
+
         return await self.repository.create_feedback(new_feedback, user.email)
 
-    async def get_verify_code(self, email: str) -> None:
+    async def get_registration_verify_code(self, email: str) -> None:
         potential_user = await self.repository.get_user_by_email(email)
         if potential_user is not None:
             raise EmailExistsException()
 
-        try:
-            code = generate_confirmation_mode()
-            task_send_to_email.delay(email, code)
-            potential_code = await self.repository.get_verify_code_by_email(email)
-            if potential_code is not None:
-                await self.repository.update_verify_code(email, code)
-            else:
-                await self.repository.create_verify_code(email, code)
+        code = self.generate_confirmation_mode()
+        message = messages.LETTER_CONFIRMATION_MESSAGE.format(code=code)
+        potential_code = await self.repository.get_verify_code_by_email(email)
+        if potential_code is not None:
+            await self.repository.update_verify_code(email, code, CodeType.for_registration)
+        else:
+            await self.repository.create_verify_code(email, code, CodeType.for_registration)
 
-        except smtplib.SMTPRecipientsRefused as e:
-            raise IncorrectEmailAddressException()
-        except Exception as e:
+        try:
+            task_send_to_email.delay(
+                subject=messages.LETTER_REGISTRATION_TITLE,
+                body=message,
+                address=email
+            )
+        except Exception:
             raise EmailSenderException()
 
-    async def check_verify_code(self, email: str, code: int) -> bool:
+    async def get_edit_password_verify_code(self, user: User) -> None:
+        code = self.generate_confirmation_mode()
+        message = messages.LETTER_CONFIRMATION_MESSAGE.format(code=code)
+        potential_code = await self.repository.get_verify_code_by_email(user.email)
+        if potential_code is not None:
+            await self.repository.update_verify_code(user.email, code, CodeType.for_reset_password)
+        else:
+            await self.repository.create_verify_code(user.email, code, CodeType.for_reset_password)
+
+        try:
+            task_send_to_email.delay(
+                subject=messages.LETTER_PASSWORD_RESET_TITLE,
+                body=message,
+                address=user.email
+            )
+        except Exception:
+            raise EmailSenderException()
+
+    async def check_verify_code(self, email: str, code: int, excepted_type: CodeType) -> bool:
         verify_code = await self.repository.get_verify_code_by_email(email)
+
+        if verify_code.type != excepted_type:
+            raise CodeTypeException(verify_code.type.value, excepted_type.value)
         if verify_code is None:
             raise IncorrectEmailAddressException()
-
         if verify_code.code != code:
             raise IncorrectVerifyCodeException()
 
         await self.repository.delete_verify_code_by_id(verify_code.id)
+
         return True
 
     async def get_user_statistic(self, user: User) -> List[DayStatistic]:
@@ -253,9 +302,9 @@ class UserService:
         from src.ai_chat.services import AIChatService
         from src.ai_chat.models import MessageBelong
 
-        welcome_chat = await AIChatService().create_new_chat(new_user, welcome_messages.WELCOME_CHAT_NAME)
+        welcome_chat = await AIChatService().create_new_chat(new_user, messages.WELCOME_CHAT_NAME)
         await AIChatService().create_new_message(
-            new_user, welcome_messages.WELCOME_CHAT_MESSAGE, welcome_chat.id, MessageBelong.assistant_message
+            new_user, messages.WELCOME_CHAT_MESSAGE, welcome_chat.id, MessageBelong.assistant_message
         )
 
         return new_user
@@ -284,8 +333,8 @@ class UserService:
             raise UserNotFoundException()
         return user
 
-    async def update_user_plan(self, user_id: uuid.UUID, plan: Plans) -> User:
-        return await self.repository.update_user_plan(user_id, plan)
+    async def update_user_plan(self, user_id: uuid.UUID, plan: Plans, plan_expire_date: datetime.date) -> User:
+        return await self.repository.update_user_plan(user_id, plan, plan_expire_date)
 
     async def change_verified_status(self, user_id: uuid.UUID) -> User:
         user = await self.get_user_by_id(user_id)
